+++ src/test/extension.test.ts
@@ -1,15 +1,28 @@
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
 import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
+import * as fs from 'fs';
+import * as path from 'path';
 
-suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+suite('Extension Test Suite', function () {
+    this.timeout(10000); 
 
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
+    vscode.window.showInformationMessage('Start all tests.');
+
+    const testFilesPath = path.join(__dirname, '../../../test-scan-files');
+    const files = fs.readdirSync(testFilesPath);
+
+    files.forEach(file => {
+        test(`Test for file: ${file}`, async () => {
+            const filePath = path.join(testFilesPath, file);
+            const document = await vscode.workspace.openTextDocument(filePath);
+            await vscode.window.showTextDocument(document);
+
+            assert.strictEqual(document.languageId, 'xml', `File ${file} should be an XML file`);
+
+            await vscode.commands.executeCommand('nmap-peek.visualize');
+
+            await new Promise(resolve => setTimeout(resolve, 5000)); // Delay for 5000 milliseconds
+            await vscode.commands.executeCommand('workbench.action.closeAllEditors');
+        });
+    });
 });